requirements/dev.txt
mysqlclient==2.0.3
requests==2.25.1
fastapi==0.63.0
uvicorn==0.13.3
SQLAlchemy==1.3.23
alembic==1.5.4
Pillow==8.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.2.0
python-multipart==0.0.5